# Add support for higher order DG on GPU