@mwarusz 2020-04-24
The new kernels are slightly slower. For a big `50^3` elements isentropic vortex simulation the profiler results are:

Old kernels
```
==5124== Profiling application: /home/mwarusz/julia-1.4.1/bin/julia --project=. test/Numerics/DGmethods/Euler/isentropicvortex.jl
==5124== Profiling result:
   Start  Duration            Grid Size      Block Size     Regs*    SSMem*    DSMem*      Size  Throughput  SrcMemType  DstMemType           Device   Context    Stream  Name
155.607s  3.6314ms         (125000 1 1)       (125 1 1)        56  4.1016KB        0B         -           -           -           -  Tesla V100-SXM2         1        17  ptxcall_gpu_volumeviscterms__7 [252]
171.982s  5.6701ms         (125000 1 1)       (125 1 1)       120  14.852KB        0B         -           -           -           -  Tesla V100-SXM2         1        19  ptxcall_gpu_volumerhs__9 [318]
```

New kernels

```
==4889== Profiling result:
   Start  Duration            Grid Size      Block Size     Regs*    SSMem*    DSMem*      Size  Throughput  SrcMemType  DstMemType           Device   Context    Stream  Name
153.394s  3.7525ms         (125000 1 1)        (25 1 1)       246  0.9766KB        0B         -           -           -           -  Tesla V100-SXM2         1        17  ptxcall_gpu_volumeviscterms__7 [252]
169.849s  5.9974ms         (125000 1 1)        (25 1 1)       250  2.1641KB        0B         -           -           -           -  Tesla V100-SXM2         1        19  ptxcall_gpu_volumerhs__9 [318]
```

Just for fun, convergence rates for the isentropic vortex with polynomial order 22:
```
┌ Info: Configuration
│ ArrayType     = CuArrays.CuArray
│ FT        = Float64
│ NumericalFlux = Rusanov
│ polynomial order = 22
└ dims          = 3
┌ Info: Starting refinement level 1
│ numelems  = (2, 2, 1)
│ dt        = 2.9629629629629629e-07
└ norm(Q₀)  = 1.1206420576248761e+03
┌ Info: Finished refinement level 1
│ norm(Q)                 = 1.1206420577876918e+03
│ norm(Q) / norm(Q₀)      = 1.0000000001452878e+00
│ norm(Q) - norm(Q₀)      = 1.6281569514831062e-07
│ norm(Q - Qe)            = 1.2633686552812546e-02
└ norm(Q - Qe) / norm(Qe) = 1.1273614496387784e-05
┌ Info: Starting refinement level 2
│ numelems  = (4, 4, 1)
│ dt        = 1.4847809948032666e-07
└ norm(Q₀)  = 1.1206420576126263e+03
┌ Info: Finished refinement level 2"""
│ norm(Q)                 = 1.1206420576126286e+03
│ norm(Q) / norm(Q₀)      = 1.0000000000000020e+00
│ norm(Q) - norm(Q₀)      = 2.2737367544323206e-12
│ norm(Q - Qe)            = 1.9999539606488015e-05
└ norm(Q - Qe) / norm(Qe) = 1.7846501004160299e-08
┌ Info: Starting refinement level 3
│ numelems  = (8, 8, 1)
│ dt        = 7.4321813452248234e-08
└ norm(Q₀)  = 1.1206420576126263e+03
┌ Info: Finished refinement level 3
│ norm(Q)                 = 1.1206420576126302e+03
│ norm(Q) / norm(Q₀)      = 1.0000000000000036e+00
│ norm(Q) - norm(Q₀)      = 3.8653524825349450e-12
│ norm(Q - Qe)            = 1.3713771197821584e-10
└ norm(Q - Qe) / norm(Qe) = 1.2237423274151325e-13
┌ Info: Convergence rates
│ rate for levels 1 → 2 = 9.303093179307787
└ rate for levels 2 → 3 = 17.153981906259983
```
The best convergence rate that I managed to get with `Float64` is 17.