@mwarusz 2020-04-24
> > The new kernels are slightly slower.
> 
> Was this expected, e.g., did we see a similar performance hit in straight CUDA testing? (I don't remember since this was mainly you and @lcw working on that.)

I did not expect this, but all the comparisons to CUDA that we did had kernels with just Euler fluxes in them, without any diffusion. It seems that diffusion increases register pressure substantially (we are near the 255 register limit even with constant viscosity) and is the main reason the new kernels are a bit slower.