@lcw 2020-04-24
> > The new kernels are slightly slower.
> 
> Was this expected, e.g., did we see a similar performance hit in straight CUDA testing? (I don't remember since this was mainly you and @lcw working on that.)

This was expected.  I saw the same in CUDA with `nvcc`.  Actually, without diffusion the Julia version of the sliced kernel is faster than the `nvcc` version.