@jkozdon 2020-04-30
So BOMEX works if I change [line 161 of bomex](https://github.com/climate-machine/CLIMA/blob/5af406aaa41110f2cc4b773b98c62396ca4f3d7f/experiments/AtmosLES/bomex.jl#L161) to
```
β_sponge = α_max * sinpi(r / 2)^2
```
instead of
```
β_sponge = α_max * sinpi(r / 2)^s.γ
```

This is with Julia 1.3 on a single V100.