@jkozdon 2020-04-30
Other useful information:

- bomex runs as is with `Float64`
- dycoms breaks with `Float32`
- in both breaking cases if the sponge source is removed the code runs